initial pass at some vanilla mob pokemob compatiblity
This is the initial pass at making sure that any random mob can be made into a pokemob.

Experimental support is added for almost all vanilla mobs, disabled by default in configs.

This gives a sample of how to go about allowing any other mod's mobs to also be pokemobs