lets village gossip reputation affect trainer sight
this is to address #598

2 configs are added:
`trainer_min_rep`
`trainer_max_rep`

these are then used, along with the gossip reputation to adjust the sight range for battle.

Winning a battle will increase reputation with that trainer by 20, losing to the trainer will also increase reputation, but by 10, this is to allow them to stop attacking you if you lose too many times.

Reputation can be lost the same way as for vanilla villages, but harming them, killing them, etc