fully convert move recipes over to a vanilla compatible format
They will still be able to load from the pokecube datapacks as well, as that allows multiple recipes in a single file, but this will now also support vanilla-style datapacks with the recipes, which may support better server-client syncs later on.

Note: Recipes added vanilla-way, like the icebeam_clay above do not currently work in JEI, this will need to be fixed later.