initial work on implementing new way to define mob spawns
This should help with grouping the pokemobs, as we can have more organised spawns. With this, all of the "forest" like spawns in one spot, and easier to add/remove and rebalance the spawn groupings.

This also uses the spawn presets now, which is part of the work for helping assist with issues such as #833