Npc overhaul part 1
First part of the NPC overhaul.

This allows "none" profession npcs and trainers to take on villager professions. When they do this, they will do the same jobs as the villagers.

The npcs can also reproduce, following similar rules to villagers.

Also updates the beach village structures to add the structure voids