some work regarding #859
This makes it so that if you are wearing one of the belts, you can store pokecubes on it. This requires assigning a hotkey for the belt slot, and pressing the hotkey will open the gui. This should still maintain balance in combat, as you need to open the gui to take things out of it.

Also made it so that if a trainer is wearing a belt, it will render the pokecubes on their belt.