next step of supporting bbmodel animations
included is a python script which generates the xml section needed for the animations, by taking them out of the model, presently it prefixes the name with `converted_`.

Added a java version of this as well, which is now used internally to load the bbmodel animations, they both presently have bugs with animations that change part locations, as well as with animations that change scale, however for rotation only ones they seem to work fine.