refactor of evolution conditions checking
They now go through a set of nested `PokemobCondition` objects, which verify if the pokemob is valid. The `HasHeldItem` specifically also checks `getEvolutionStack` from the `IPokemob`.

More `PokemobCondition` can be registered, and they load from json elements, and are effectively a predicate on a `IPokemob`.

Implemented `PokemobCondition` types for the things which were used by all of the existing evolutions.