refactor of recalling on death code
this refactors the code for where death recalling/reviving is handled.
The new system is handled via events rather than in the EntityPokemob
this adjustment also allows for non-EntityPokemob pokemobs to properly recall to their pokecubes when they die

Pokemobs now recall to their cubes immediately, without waiting to play the death animation. The mob is still left in the world to play the animation client side, and the cube won't allow tossing until the death timer cooldown has expired

This does then remove the ability to revive your own pokemob by quickly right clicking it with a revive.