fixes wild trainer spawns not occuring
also involves a cleanup of the data format for specifiying their spawns. It is now just a standard spawn rule with an additional "rate" field, rather than rate needing to be wrapped in a "values" map

This addresses the rest of the warnings about old spawn rule formats as well.