Harvest items compatibility changes
Modified the IHarvester to have an isAvailable function and then used
it for checking if the item was gatherable in hasStuff. This allows
mod compatibility to write a new IHarvester and register it instead of
being blocked at the hasStuff, as if the item is listed the
harvest_extra tag, that will just use the default harvest function
which may be undesirable.

Moved the drop list to be retrieved before the block is broken, as in
some cases the block will drop nothing, likely because the block will
be air.